unfollow 구현

+++ src/main/java/com/hanghae99/finalproject/model/repository/FollowRepository.java
@@ -4,7 +4,7 @@ import com.hanghae99.finalproject.model.entity.Follow;
 import com.hanghae99.finalproject.model.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FollowRepository  extends JpaRepository<Follow, Integer> {
+public interface FollowRepository  extends JpaRepository<Follow, Long> {
 
     Follow findFollowByFollowingAndFollower(Users following_id, Users follower_id);
